requirements.txt
flask==1.1.2
spacy==2.3.2
pandas>=1.0.5
nltk
imdbpy